requirements.txt
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
pycodestyle==2.6.0
pylint==2.6.0
pymongo==3.11.0
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
